Move CorsEndpointProperties to the upper package
`CorsEndpointProperties` lives in `org.springframework.boot.actuate.autoconfigure.endpoint.web.servlet` but is also used in `org.springframework.boot.actuate.autoconfigure.endpoint.web.reactive`, so this PR moves it to its common parent package.

This commit also extracts `CorsConfiguration` creation logic duplicated in `WebMvcEndpointManagementContextConfiguration` and `WebFluxEndpointManagementContextConfiguration` into `CorsEndpointProperties`.
